Passage en Integer de mes entity au lieu de Long

+++ src/main/java/com/site/digitalBook/repository/CategorieRepository.java
@@ -3,6 +3,6 @@ package com.site.digitalBook.repository;
 import com.site.digitalBook.entity.Categorie;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CategorieRepository extends JpaRepository<Categorie, Long> {
+public interface CategorieRepository extends JpaRepository<Categorie, Integer> {
 }
